backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tqdm==4.66.5
beautifulsoup4==4.12.3
requests==2.32.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
gspread==6.1.4
oauth2client==4.1.3
pydantic==2.9.2